refactor: router 디렉토리 파일 이름 규칙 적용, 라우터 정리
