+++ src/router/DetailPageRoutes.js
